+++ bzfs_main/argparse_cli.py
@@ -20,7 +20,6 @@ from __future__ import (
 import argparse
 import dataclasses
 import itertools
-import platform
 
 from bzfs_main.argparse_actions import (
     CheckPercentRange,
@@ -158,7 +157,7 @@ source or destination host. Only the underlying 'zfs' CLI must be installed on b
 ZFS permissions by administrators via 'zfs allow' delegation mechanism.
 
 {PROG_NAME} is written in Python and continuously runs a wide set of unit tests and integration tests to ensure
-coverage and compatibility with old and new versions of ZFS on Linux, FreeBSD and Solaris, on all Python
+coverage and compatibility with old and new versions of ZFS on Linux and FreeBSD, on all Python
 versions ≥ 3.9 (including latest stable which is currently python-3.13).
 
 {PROG_NAME} is a stand-alone program with zero required dependencies, akin to a
@@ -1174,7 +1173,7 @@ as how many src snapshots and how many GB of data are missing on dst, etc.
              f"any way you like, as {PROG_NAME} (without --no-use-bookmark) will happily work with whatever "
              "bookmarks currently exist, if any.\n\n")
 
-    ssh_cipher_default = "^aes256-gcm@openssh.com" if platform.system() != "SunOS" else ""
+    ssh_cipher_default = "^aes256-gcm@openssh.com"
     # ^aes256-gcm@openssh.com cipher: for speed with confidentiality and integrity
     # measure cipher perf like so: count=5000; for i in $(seq 1 3); do echo "iteration $i:"; for cipher in $(ssh -Q cipher); do dd if=/dev/zero bs=1M count=$count 2> /dev/null | ssh -c $cipher -p 40999 127.0.0.1 "(time -p cat) > /dev/null" 2>&1 | grep real | awk -v count=$count -v cipher=$cipher '{print cipher ": " count / $2 " MB/s"}'; done; done
     # see https://gbe0.com/posts/linux/server/benchmark-ssh-ciphers/